Reorganize and recategorize modules
