+++ src/report/logging.rs
